feat: Enhance VerifikasiSuratKuasaController with approval and rejection logic
- Updated approve method to include decryption of ID, duplicate checks, and transaction handling.
- Added job dispatch for PDF generation after successful approval.
- Improved logging for approval and rejection actions.
- Changed relationship from belongsTo to hasOne in PendaftaranSuratKuasaModel for register.
- Adjusted RegisterSuratKuasaModel to specify foreign key in belongsTo relationship.
- Updated data-surat-kuasa view to restrict delete functionality based on user role.
- Modified detail-surat-kuasa view to display correct panitera name.
- Enhanced PDF barcode template with dynamic data and QR code generation.
- Added new route for verifying surat kuasa by UUID.
- Created GenerateBarcodeSuratKuasaPDF job for handling PDF generation asynchronously.

+++ resources/views/admin/surat-kuasa/data-surat-kuasa.blade.php
@@ -43,41 +43,43 @@
 @endsection
 @push('scripts')
     {!! $dataTable->scripts(attributes: ['type' => 'module']) !!}
-    <script type="text/javascript">
-        window.deleteData = async function(url) {
-            const result = await Swal.fire({
-                title: 'Apakah Anda yakin?',
-                text: "Data yang akan dihapus tidak dapat dikembalikan!",
-                icon: 'warning',
-                showCancelButton: true,
-                confirmButtonColor: '#3085d6',
-                cancelButtonColor: '#d33',
-                confirmButtonText: 'Ya, hapus!',
-                cancelButtonText: 'Batal'
-            });
+    @if (Auth::user()->role != \App\Enum\RoleEnum::User->value)
+        <script type="text/javascript">
+            window.deleteData = async function(url) {
+                const result = await Swal.fire({
+                    title: 'Apakah Anda yakin?',
+                    text: "Data yang akan dihapus tidak dapat dikembalikan!",
+                    icon: 'warning',
+                    showCancelButton: true,
+                    confirmButtonColor: '#3085d6',
+                    cancelButtonColor: '#d33',
+                    confirmButtonText: 'Ya, hapus!',
+                    cancelButtonText: 'Batal'
+                });
 
-            if (result.isConfirmed) {
-                const csrfToken = document.querySelector('meta[name="csrf-token"]').getAttribute('content');
-                try {
-                    const response = await fetch(url, {
-                        method: 'DELETE',
-                        headers: {
-                            'X-CSRF-TOKEN': csrfToken,
-                            'Content-Type': 'application/json'
+                if (result.isConfirmed) {
+                    const csrfToken = document.querySelector('meta[name="csrf-token"]').getAttribute('content');
+                    try {
+                        const response = await fetch(url, {
+                            method: 'DELETE',
+                            headers: {
+                                'X-CSRF-TOKEN': csrfToken,
+                                'Content-Type': 'application/json'
+                            }
+                        });
+                        const data = await response.json();
+                        if (data.success) {
+                            await Swal.fire('Berhasil!', data.message, 'success');
+                            window.LaravelDataTables['pendaftaransuratkuasa-table'].ajax.reload();
+                        } else {
+                            Swal.fire('Gagal!', data.message || 'Terjadi kesalahan saat menghapus data.', 'error');
                         }
-                    });
-                    const data = await response.json();
-                    if (data.success) {
-                        await Swal.fire('Berhasil!', data.message, 'success');
-                        window.LaravelDataTables['panitera-table'].ajax.reload();
-                    } else {
-                        Swal.fire('Gagal!', data.message || 'Terjadi kesalahan saat menghapus data.', 'error');
+                    } catch (error) {
+                        console.error('Error:', error);
+                        Swal.fire('Oops', 'Terjadi kesalahan.', 'error');
                     }
-                } catch (error) {
-                    console.error('Error:', error);
-                    Swal.fire('Error!', 'Terjadi kesalahan.', 'error');
                 }
             }
-        }
-    </script>
+        </script>
+    @endif
 @endpush
